app/doormanager/requirements.txt
certifi
paho-mqtt
PyYAML
RPi.Gpio
gpiozero
fastapi
fastapi-mqtt
fastapi-utils
gunicorn
uvicorn[standard]
# PyDispatcher
pydantic
# psycopg2
psycopg2-binary
# fastapi-sqlalchemy
sqlalchemy
# databases[postgresql]
python-box
dependency-injector